version 1.0
Significant changes to the wireless configuration.  Removed all "hard coding" of host and directory and also added a backup server.  Overall made the wireless set up much more user friendly.  Also optimize the size of string that is sent as I was approaching the limit and could exceed the limit causing home page not to load.  Added changes to allow EVSE to run "headless" without any built in display and having all control and data sent via browser.  Made significant changes and may not have marked everything with a //bhc